Backend/app/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
google-generativeai
langchain
sentence-transformers
chromadb
PyMuPDF
pytesseract
Pillow